refactored codelist into separate file
