+++ rust/codelist-rs/src/lib.rs
@@ -5,3 +5,4 @@ pub mod types;
 pub mod code_entry;
 pub mod metadata;
 pub mod codelist_factory;
+pub mod codelist_options;
